move ZeroEnergyAndCO2e to other
